Move Tasks and Estimation Procedure to a separate submodule
